+++ examples/blocks/src/market/market.js
@@ -10,6 +10,10 @@
 // ┃ of the [Apache License 2.0](https://www.apache.org/licenses/LICENSE-2.0). ┃
 // ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
 
+import "/node_modules/@finos/perspective-viewer/dist/cdn/perspective-viewer.js";
+import "/node_modules/@finos/perspective-viewer-datagrid/dist/cdn/perspective-viewer-datagrid.js";
+import "/node_modules/@finos/perspective-viewer-d3fc/dist/cdn/perspective-viewer-d3fc.js";
+
 import perspective from "/node_modules/@finos/perspective/dist/cdn/perspective.js";
 
 const MSG_BATCH_TIMEOUT = 50;