+++ packages/perspective-workspace/src/ts/utils/observable_map.ts
@@ -10,60 +10,34 @@
 // ┃ of the [Apache License 2.0](https://www.apache.org/licenses/LICENSE-2.0). ┃
 // ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
 
-import { h } from "@lumino/virtualdom/src";
-import { Menu } from "@lumino/widgets/src/menu";
+export class ObservableMap<K, V> extends Map<K, V> {
+    _set_listener?: (name: K, val: V) => void;
+    _delete_listener?: (name: K) => void;
 
-export class MenuRenderer extends Menu.Renderer {
-    formatLabel(data) {
-        let { label, mnemonic } = data.item;
-        if (mnemonic < 0 || mnemonic >= label.length) {
-            return label;
+    set(name: K, item: V) {
+        this._set_listener?.(name, item);
+        super.set(name, item);
+        return this;
+    }
+
+    get(name: K): V {
+        return super.get(name)!;
+    }
+
+    delete(name: K) {
+        const result = this._delete_listener?.(name);
+        if (result) {
+            return super.delete(name);
+        } else {
+            return false;
         }
-
-        let prefix = label.slice(0, mnemonic);
-        let suffix = label.slice(mnemonic + 1);
-        let char = label[mnemonic];
-        let span = h.span(
-            {
-                className: "lm-Menu-itemMnemonic p-Menu-itemMnemonic",
-            },
-            char
-        );
-
-        return [prefix, span, suffix];
     }
 
-    renderLabel(data) {
-        let content = this.formatLabel(data);
-        return h.div(
-            {
-                className: "lm-Menu-itemLabel p-Menu-itemLabel",
-            },
-            content
-        );
+    addSetListener(listener: (name: K, val: V) => void) {
+        this._set_listener = listener;
     }
 
-    renderSubmenu(data) {
-        return h.div({
-            className: "lm-Menu-itemSubmenuIcon" + " p-Menu-itemSubmenuIcon",
-        });
-    }
-
-    renderItem(data) {
-        let className = this.createItemClass(data);
-        let dataset = this.createItemDataset(data);
-        let aria = this.createItemARIA(data);
-        return h.li(
-            {
-                className,
-                dataset,
-                tabindex: "0",
-                onfocus: data.onfocus,
-                ...aria,
-            },
-            this.renderLabel(data),
-            this.renderShortcut(data),
-            this.renderSubmenu(data)
-        );
+    addDeleteListener(listener: (name: K) => void) {
+        this._delete_listener = listener;
     }
 }